Undo cells and empty cells
Added JS that allows the ability to mark empty cells white, and also undo actions, by turning them back to gold. Made it so that upon clicking the show solution button, the puzzle solution will appear. Added the start of loops that will mark incorrect cells in different colors when the peek button is clicked.